backend/requirements.txt
attrs==21.2.0
blinker==1.4
certifi==2021.5.30
charset-normalizer==2.0.3
click==6.6
Flask==0.12.1
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.1
greenlet==1.1.0
idna==3.2
importlib-metadata==4.5.0
iniconfig==1.1.1
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.1.0
packaging==21.0
passlib==1.7.4
Pillow==8.0.0
pkg-resources==0.0.0
pluggy==0.13.1
py==1.10.0
pyparsing==2.4.7
pytest==6.2.4
requests==2.26.0
SQLAlchemy==1.3.23
toml==0.10.2
typing-extensions==3.7.2
urllib3==1.26.6
Werkzeug==0.11.15
zipp==3.4.1
